// frontend/src/types/Donations.ts
export enum Donations {
  forest = 'Forest',
  charity = 'Charity',
  seas = 'Seas',
  agriculture = 'Agriculture',
  education = 'Education',
}

export enum DonationBackgroundPath {
  forest = '/background/Forest.jpg',
  charity = '/background/Charity.jpg',
  education = '/background/Education.jpg',
  seas = '/background/Sea.jpg',
  agriculture = '/background/Agriculture.jpg',
}